pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-model"
version = "0.1.0"
description = "Multi-modal wrapper for Stable Diffusion models and more – image, video and text manipulation toolkit for low-power devices."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "accelerate>=0.26.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
video = ["opencv-python>=4.8.0"]
memory = ["psutil>=5.9.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "opencv-python>=4.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_model*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
